Allow sorting dynamic columns by more than just numbers
